OCT-111: moved GetProductIndetifiersQuery
