+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/Product/GetProductIdentifiersQueryTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Catalog\Product;
 
 use Akeneo\Catalogs\Domain\Operator;
-use Akeneo\Catalogs\Infrastructure\Persistence\GetProductIdentifiersQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product\GetProductIdentifiersQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetEnabled;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetTextValue;
